refactor: Use custom mergeWith at index as well
Related to #134.
