+++ src/index.ts
@@ -1,5 +1,5 @@
-import { mergeWith } from "lodash";
 import { Configuration } from "webpack";
+import mergeWith from "./merge-with";
 import joinArrays from "./join-arrays";
 import unique from "./unique";
 import {
@@ -44,15 +44,14 @@ function mergeWithCustomize(options: ICustomizeOptions) {
   ): Configuration {
     if (configurations.length === 0) {
       if (Array.isArray(firstConfiguration)) {
-        return mergeWith({}, ...firstConfiguration, joinArrays(options));
+        return mergeWith(firstConfiguration, joinArrays(options));
       }
 
       return firstConfiguration;
     }
 
     return mergeWith(
-      {},
-      ...[firstConfiguration].concat(configurations),
+      [firstConfiguration].concat(configurations),
       joinArrays(options)
     );
   };
@@ -62,11 +61,11 @@ function customizeObject(rules: ICustomizeRules) {
   return (a: any, b: any, key: Key) => {
     switch (rules[key]) {
       case CustomizeRule.Prepend:
-        return mergeWith({}, b, a, joinArrays());
+        return mergeWith([b, a], joinArrays());
       case CustomizeRule.Replace:
         return b;
       case CustomizeRule.Append:
-        return mergeWith({}, a, b, joinArrays());
+        return mergeWith([a, b], joinArrays());
     }
   };
 }
